+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Adds a `Add GitHub Repository to Workspace...` command (`remotehub.addRepository`) which allows you to search for a remote GitHub repository to open in the current workspace
+- Adds a *go back* item to the repository search quick pick menu
+
+### Changed
+- Changes the `Open GitHub Repository...` command (`remotehub.openRepository`) which allows you to search for a remote GitHub repository to open, replacing the current workspace &mdash; eventually this will open in a new window, once I figure out a good way to do it
+
+### Removed
+- Removes the `Open GitHub Repository By Owner...` command (`remotehub.openRepositoryByOwner`) as it is functionality has been included in the `Open GitHub Repository...` command
+
 ## [0.1.1] - 2018-06-06
 ### Fixed
 - Fixes incorrect setting name in README